feat: profile click inside chats section
