+++ src/modules/chats/components/chatDetails/index.tsx
@@ -1,4 +1,3 @@
-import React, { useEffect } from 'react';
 import InfiniteScroll from 'react-infinite-scroller';
 import { SendMessage } from 'src/modules/chats/components/sendMessage';
 import { AvatarLabelGroup } from 'src/modules/general/components/avatarLabelGroup';
@@ -16,26 +15,12 @@ export const ChatDetails: React.FC<ChatDetailsProps> = ({
   newSocketMessage,
   setChats,
 }) => {
-  const { messages, onSend, account, loadMore, hasMore, page, setMessages, onAvatarClick } = useChatDetails(
+  const { sortedMessages, onSend, account, loadMore, hasMore, onAvatarClick } = useChatDetails(
     selectedChatId,
     chats,
     setChats,
+    newSocketMessage,
   );
-
-  const sorted = messages?.sort((a, b) => (new Date(a.created_at) > new Date(b.created_at) ? 1 : -1));
-
-  useEffect(() => {
-    if (page === 1) {
-      const messageBody = document.getElementById('chat-list-div');
-      if (messageBody) messageBody.scrollTop = messageBody.scrollHeight - messageBody.clientHeight;
-    }
-  }, [messages]);
-
-  useEffect(() => {
-    if (newSocketMessage && !messages.find(item => item.id === newSocketMessage.id)) {
-      setMessages([...messages, newSocketMessage]);
-    }
-  }, [newSocketMessage]);
 
   return (
     <div className="flex flex-col w-full h-full">
@@ -62,13 +47,14 @@ export const ChatDetails: React.FC<ChatDetailsProps> = ({
           isReverse
           className="flex-1"
         >
-          {sorted?.map((item, index) => (
+          {sortedMessages?.map(message => (
             <ChatDetailItem
-              key={item.id}
-              message={item}
+              key={message.id}
+              message={message}
               senderAvatar={account.img}
               senderType={account.type}
               senderName={account.name}
+              onAvatarClick={onAvatarClick}
             />
           ))}
         </InfiniteScroll>
